Changing from 16 to 64 bit for number of elements in an array

+++ include/ndarray/ndarray_initializers.i.hpp
@@ -49,7 +49,7 @@ template <class T>
 ndarray<T> ndarray<T>::fromarray(std::vector<uint16_t> shape, T *input)
 {
     // Multiplies all elements in lengths together
-    uint16_t values_length = 1;
+    uint64_t values_length = 1;
     for (auto i : shape)
         values_length *= i;
     return ndarray<T>(shape, {input, input + values_length});
